+++ scout/commands/load/user.py
@@ -1,27 +1,12 @@
-#!/usr/bin/env python
-# encoding: utf-8
 import logging
-from typing import Optional
 
 import click
 from flask.cli import with_appcontext
 
-from scout.build.user import build_user
+from scout.load.user import save_user
 from scout.server.extensions import store
 
 LOG = logging.getLogger(__name__)
-
-
-def save_user(user_info: dict) -> Optional[dict]:
-    """Saves a new user document to the database, or raises an error if the operation fails."""
-
-    user_obj: dict = build_user(user_info)
-    try:
-        new_user: dict = store.add_user(user_obj)
-        return new_user
-    except Exception as err:
-        LOG.warning(err)
-        raise click.Abort()
 
 
 @click.command("user", short_help="Load a user")
